requirements/base.txt
Django==4.2.5
celery==5.3.4
django-cors-headers==4.2.0
django-environ==0.11.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Faker==19.6.2
psycopg2==2.9.8
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
django-jalali-date==1.0.2
django-lifecycle==1.0.2
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.9.1
Pillow==10.0.1
